# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rapid ERP Retail API",
        "description": "API-only, multi-tenant retail ERP. Every response uses the standard envelope: {status, code, message, data}.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate a user (issues a token, or a login OTP for 2FA users)",
                "description": "Authenticate a user. Returns an access token directly, unless the user\nrequires verification — then a login OTP is emailed and the client must\ncall verify-otp to obtain the token.",
                "operationId": "9d0bb5c770cbec19b0b9566606ef4adb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "superadmin@boilerplate.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Either logged in (token returned) or an OTP was sent for verification.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccess"
                                        }
                                    ]
                                },
                                "examples": {
                                    "token": {
                                        "summary": "Standard login (verification not required)",
                                        "value": {
                                            "status": "success",
                                            "code": 200,
                                            "message": "Logged in successfully.",
                                            "data": {
                                                "user": {
                                                    "id": 1,
                                                    "name": "Super Administrator",
                                                    "email": "superadmin@boilerplate.test",
                                                    "phone": "+10000000000",
                                                    "email_verified_at": "04-June-2026 15:01:33"
                                                },
                                                "token_type": "Bearer",
                                                "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                                            }
                                        }
                                    },
                                    "otp": {
                                        "summary": "Verification required (OTP emailed)",
                                        "value": {
                                            "status": "success",
                                            "code": 200,
                                            "message": "A verification code has been sent to your email.",
                                            "data": {
                                                "requires_verification": true
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    ]
                                },
                                "example": {
                                    "status": "failed",
                                    "code": 401,
                                    "message": "Invalid credentials.",
                                    "data": {
                                        "errors": []
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
        "/api/v1/auth/login/verify-otp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify a login OTP and issue an access token",
                "description": "Verify a login OTP and issue an access token.",
                "operationId": "ebc2cb945ba7361eafba29dc81d06eb8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "superadmin@boilerplate.test"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccess"
                                        }
                                    ]
                                },
                                "example": {
                                    "status": "success",
                                    "code": 200,
                                    "message": "Logged in successfully.",
                                    "data": {
                                        "user": {
                                            "id": 1,
                                            "name": "Super Administrator",
                                            "email": "superadmin@boilerplate.test",
                                            "phone": "+10000000000",
                                            "email_verified_at": "04-June-2026 15:01:33"
                                        },
                                        "token_type": "Bearer",
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
        "/api/v1/auth/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Request a password-reset OTP",
                "description": "Issue a password-reset OTP to the user's email.",
                "operationId": "4011ca27431b5cd53a225688e531de47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "superadmin@boilerplate.test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccess"
                                        }
                                    ]
                                },
                                "example": {
                                    "status": "success",
                                    "code": 200,
                                    "message": "A password reset OTP has been sent to your email.",
                                    "data": null
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
        "/api/v1/auth/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset password using an OTP",
                "description": "Reset the password using a valid OTP.",
                "operationId": "aba7a33c4064b5b1a2f3a8a70c20094b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "superadmin@boilerplate.test"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new-secret-password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new-secret-password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccess"
                                        }
                                    ]
                                },
                                "example": {
                                    "status": "success",
                                    "code": 200,
                                    "message": "Your password has been reset successfully.",
                                    "data": null
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Return the currently authenticated user",
                "description": "Return the currently authenticated user.",
                "operationId": "7edb820901b8b3a880fc4d7a0a7290fb",
                "responses": {
                    "200": {
                        "description": "Authenticated user retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccess"
                                        }
                                    ]
                                },
                                "example": {
                                    "status": "success",
                                    "code": 200,
                                    "message": "Authenticated user retrieved.",
                                    "data": {
                                        "id": 1,
                                        "name": "Super Administrator",
                                        "email": "superadmin@boilerplate.test",
                                        "phone": "+10000000000",
                                        "email_verified_at": "04-June-2026 15:01:33"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Revoke the current access token",
                "description": "Revoke the current access token.",
                "operationId": "3a27a9e02ceb521eb3f98c8d2394aef9",
                "responses": {
                    "200": {
                        "description": "Logged out successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccess"
                                        }
                                    ]
                                },
                                "example": {
                                    "status": "success",
                                    "code": 200,
                                    "message": "Logged out successfully.",
                                    "data": null
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiSuccess": {
                "title": "Success envelope",
                "description": "Standard success envelope.",
                "properties": {
                    "status": {
                        "description": "Central, reusable OpenAPI components that mirror the response envelope built\nby {@see JsonResponseService}. Endpoints reference these so the\nSwagger docs always match the actual JSON contract.",
                        "type": "string",
                        "example": "success"
                    },
                    "code": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful."
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiError": {
                "title": "Error envelope",
                "description": "Standard failed envelope. Detail messages live in data.errors (list of strings).",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "failed"
                    },
                    "code": {
                        "type": "integer",
                        "example": 500
                    },
                    "message": {
                        "type": "string",
                        "example": "Something went wrong."
                    },
                    "data": {
                        "properties": {
                            "errors": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "Unauthorized": {
                "description": "Authentication required / token missing or invalid.",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            ]
                        },
                        "example": {
                            "status": "failed",
                            "code": 401,
                            "message": "Please login first.",
                            "data": {
                                "errors": []
                            }
                        }
                    }
                }
            },
...
            "ValidationError": {
                "description": "The given data was invalid.",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            ]
                        },
                        "example": {
                            "status": "failed",
                            "code": 422,
                            "message": "The given data was invalid.",
                            "data": {
                                "errors": [
                                    "The email field is required.",
                                    "The password field is required."
                                ]
                            }
                        }
                    }
                }
            },
            "TooManyRequests": {
                "description": "Rate limit exceeded.",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            ]
                        },
                        "example": {
                            "status": "failed",
                            "code": 429,
                            "message": "Too many requests.",
                            "data": {
                                "errors": []
                            }
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Passport access token. Send as: Authorization: Bearer {token}.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}